change order list

+++ order/models.py
@@ -17,7 +17,7 @@ class Order(models.Model):
 
 	title = models.CharField(max_length=50)
 	size = models.CharField(choices=SIZE_OPTIONS, max_length=1)
-	quantity = models.IntegerField()
+	quantity = models.PositiveSmallIntegerField()
 	status = models.CharField(choices=STATUS, max_length=1)
 	created_at = models.DateTimeField(auto_now_add=True)
 	delivered_at = models.DateTimeField(null=True, blank=True)
